+++ basic.go
@@ -1,7 +1,6 @@
 package chomp
 
 import (
-	"fmt"
 	"strings"
 )
 
@@ -151,63 +150,6 @@ func Until(str string) Combinator[string] {
 	}
 }
 
-// Opt allows a combinator to be optional. Any error returned by the underlying
-// combinator will be swallowed. The parsed text will not be modified if the
-// underlying combinator did not run.
-//
-//	chomp.Opt(chomp.Tag("Hey"))("Hello, World!")
-//	// ("Hello, World!", "", nil)
-func Opt[T Result](c Combinator[T]) Combinator[T] {
-	return func(s string) (string, T, error) {
-		rem, out, _ := c(s)
-		return rem, out, nil
-	}
-}
-
-// S wraps the result of the inner combinator within a string slice.
-// Combinators of differing return types can be successfully chained
-// together while using this conversion combinator.
-//
-//	chomp.S(chomp.Until(","))("Hello, World!")
-//	// (", World!", []string{"Hello"}, nil)
-func S(c Combinator[string]) Combinator[[]string] {
-	return func(s string) (string, []string, error) {
-		rem, ext, err := c(s)
-		if err != nil {
-			return rem, nil, err
-		}
-
-		return rem, []string{ext}, err
-	}
-}
-
-// I extracts and returns a single string from the result of the inner combinator.
-// Combinators of differing return types can be successfully chained together while
-// using this conversion combinator.
-//
-//	chomp.I(chomp.SepPair(
-//		chomp.Tag("Hello"),
-//		chomp.Tag(", "),
-//		chomp.Tag("World")), 1)("Hello, World!")
-//	// ("!", "World", nil)
-func I(c Combinator[[]string], i int) Combinator[string] {
-	return func(s string) (string, string, error) {
-		rem, ext, err := c(s)
-		if err != nil {
-			return rem, "", err
-		}
-
-		if i < 0 || i >= len(ext) {
-			return rem, "", ParserError{
-				Err:  fmt.Errorf("index %d is out of bounds within string slice of %d elements", i, len(ext)),
-				Type: "i",
-			}
-		}
-
-		return rem, ext[i], nil
-	}
-}
-
 // Prefixed will firstly scan the input text for a defined prefix and discard it.
 // The remaining input text will be matched against the [Combinator] and returned
 // if successful. Both combinators must match.
@@ -260,22 +202,5 @@ func Suffixed(c, suf Combinator[string]) Combinator[string] {
 func Eol() Combinator[string] {
 	return func(s string) (string, string, error) {
 		return Suffixed(WhileNotN(IsLineEnding, 0), Opt(Crlf()))(s)
-	}
-}
-
-// Peek will scan the text and apply the parser without consuming any of the input.
-// Useful if you need to lookahead.
-//
-//	chomp.Peek(chomp.Tag("Hello"))("Hello, World!")
-//	// ("Hello, World!", "Hello", nil)
-//
-//	chomp.Peek(
-//		chomp.Many(chomp.Suffixed(chomp.Tag(" "), chomp.Until(" "))),
-//	)("Hello and Good Morning!")
-//	// ("Hello and Good Morning!", []string{"Hello", "and", "Good"}, nil)
-func Peek[T Result](c Combinator[T]) Combinator[T] {
-	return func(s string) (string, T, error) {
-		_, ext, err := c(s)
-		return s, ext, err
 	}
 }
